Fixed pedantic toolchain warnings.

+++ tests/run-parser-test-suite.c
@@ -164,7 +164,7 @@ void print_escaped(yaml_char_t * str, size_t length)
     int i;
     char c;
 
-    for (i = 0; i < length; i++) {
+    for (i = 0; i < (int)length; i++) {
         c = *(str + i);
         if (c == '\\')
             printf("\\\\");
